Move type to its own file
